Fix racy mvid tests.
`DeterministicMvidWorks`, `NewMvidWorks`, and `RetainMvid` were running as part of `All.CommandLineTests` and as part of their own `IndividualTests`.

The individual test is going to use the same sandbox location as the `All.CommandLineTests` run uses.

This means that the two tests can conflict.  Which I suspect explains why our CI randomly hit the error below.

Given that trimming these tests twice is wasteful.  Let's update the individual test to run `Check`, giving the individual test the same coverage as when the test was running in `All.CommandsLineTests`.  Then lets move the tests into an `Individual` folder so that they are not picked up by `All.CommandsLineTests`.  Other tests in `IndividualTests` are following this pattern.

```
1) Failed : Mono.Linker.Tests.TestCases.All.CommandLine.Mvid.DeterministicMvidWorks
[19:40:23.176 Information] System.UnauthorizedAccessException : Access to the path 'test.exe' is denied.
   at System.IO.FileSystem.RemoveDirectoryRecursive(String fullPath, WIN32_FIND_DATA& findData, Boolean topLevel)
   at System.IO.FileSystem.RemoveDirectory(String fullPath, Boolean recursive)
   at Mono.Linker.Tests.Extensions.NPath.Delete(DeleteMode deleteMode) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\Extensions\NiceIO.cs:line 580
   at Mono.Linker.Tests.Extensions.Extensions.Delete(IEnumerable`1 self) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\Extensions\NiceIO.cs:line 889
   at Mono.Linker.Tests.Extensions.NPath.DeleteContents() in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\Extensions\NiceIO.cs:line 614
   at Mono.Linker.Tests.TestCasesRunner.TestCaseSandbox..ctor(TestCase testCase, NPath rootTemporaryDirectory, String namePrefix) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Trimming.Tests.Shared\TestCaseSandbox.cs:line 45
   at Mono.Linker.Tests.TestCasesRunner.TestRunner.Run(TestCase testCase) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Trimming.Tests.Shared\TestRunner.cs:line 35
   at Mono.Linker.Tests.TestCases.All.Run(TestCase testCase) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\TestCases\TestSuites.cs:line 304
   at Mono.Linker.Tests.TestCases.All.CommandLineTests(TestCase testCase) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\TestCases\TestSuites.cs:line 64
   at InvokeStub_All.CommandLineTests(Object, Span`1)
   at System.Reflection.RuntimeMethodInfo.Invoke(Object obj, BindingFlags invokeAttr, Binder binder, Object[] parameters, CultureInfo culture)
2) Failed : Mono.Linker.Tests.TestCases.All.CommandLine.Mvid.NewMvidWorks
System.UnauthorizedAccessException : Access to the path 'test.exe' is denied.
   at System.IO.FileSystem.RemoveDirectoryRecursive(String fullPath, WIN32_FIND_DATA& findData, Boolean topLevel)
   at System.IO.FileSystem.RemoveDirectory(String fullPath, Boolean recursive)
   at Mono.Linker.Tests.Extensions.NPath.Delete(DeleteMode deleteMode) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\Extensions\NiceIO.cs:line 580
   at Mono.Linker.Tests.Extensions.Extensions.Delete(IEnumerable`1 self) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\Extensions\NiceIO.cs:line 889
   at Mono.Linker.Tests.Extensions.NPath.DeleteContents() in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\Extensions\NiceIO.cs:line 614
   at Mono.Linker.Tests.TestCasesRunner.TestCaseSandbox..ctor(TestCase testCase, NPath rootTemporaryDirectory, String namePrefix) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Trimming.Tests.Shared\TestCaseSandbox.cs:line 45
   at Mono.Linker.Tests.TestCasesRunner.TestRunner.Run(TestCase testCase) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Trimming.Tests.Shared\TestRunner.cs:line 35
   at Mono.Linker.Tests.TestCases.All.Run(TestCase testCase) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\TestCases\TestSuites.cs:line 304
   at Mono.Linker.Tests.TestCases.All.CommandLineTests(TestCase testCase) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\TestCases\TestSuites.cs:line 64
   at InvokeStub_All.CommandLineTests(Object, Span`1)
   at System.Reflection.RuntimeMethodInfo.Invoke(Object obj, BindingFlags invokeAttr, Binder binder, Object[] parameters, CultureInfo culture)
3) Failed : Mono.Linker.Tests.TestCases.All.CommandLine.Mvid.RetainMvid
System.UnauthorizedAccessException : Access to the path 'test.exe' is denied.
   at System.IO.FileSystem.RemoveDirectoryRecursive(String fullPath, WIN32_FIND_DATA& findData, Boolean topLevel)
   at System.IO.FileSystem.RemoveDirectory(String fullPath, Boolean recursive)
   at Mono.Linker.Tests.Extensions.NPath.Delete(DeleteMode dele
[19:40:23.177 Information] teMode) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\Extensions\NiceIO.cs:line 580
   at Mono.Linker.Tests.Extensions.Extensions.Delete(IEnumerable`1 self) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\Extensions\NiceIO.cs:line 889
   at Mono.Linker.Tests.Extensions.NPath.DeleteContents() in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\Extensions\NiceIO.cs:line 614
   at Mono.Linker.Tests.TestCasesRunner.TestCaseSandbox..ctor(TestCase testCase, NPath rootTemporaryDirectory, String namePrefix) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Trimming.Tests.Shared\TestCaseSandbox.cs:line 45
   at Mono.Linker.Tests.TestCasesRunner.TestRunner.Run(TestCase testCase) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Trimming.Tests.Shared\TestRunner.cs:line 35
   at Mono.Linker.Tests.TestCases.All.Run(TestCase testCase) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\TestCases\TestSuites.cs:line 304
   at Mono.Linker.Tests.TestCases.All.CommandLineTests(TestCase testCase) in C:\build\output\unity\il2cpp\external\runtime\src\tools\illink\test\Mono.Linker.Tests\TestCases\TestSuites.cs:line 64
   at InvokeStub_All.CommandLineTests(Object, Span`1)
   at System.Reflection.RuntimeMethodInfo.Invoke(Object obj, BindingFlags invokeAttr, Binder binder, Object[] parameters, CultureInfo culture)
```
