+++ src/tools/illink/test/Mono.Linker.Tests.Cases/CommandLine/Mvid/Individual/DeterministicMvidWorks.cs
@@ -1,10 +1,10 @@
 using Mono.Linker.Tests.Cases.Expectations.Assertions;
 using Mono.Linker.Tests.Cases.Expectations.Metadata;
 
-namespace Mono.Linker.Tests.Cases.CommandLine.Mvid
+namespace Mono.Linker.Tests.Cases.CommandLine.Mvid.Individual
 {
-    [SetupLinkerArgument("--new-mvid", "false")]
-    public class RetainMvid
+    [SetupLinkerArgument("--deterministic", "true")]
+    public class DeterministicMvidWorks
     {
         public static void Main()
         {
