+++ src/tools/illink/test/Mono.Linker.Tests.Cases/CommandLine/Mvid/Individual/NewMvidWorks.cs
@@ -1,10 +1,10 @@
 using Mono.Linker.Tests.Cases.Expectations.Assertions;
 using Mono.Linker.Tests.Cases.Expectations.Metadata;
 
-namespace Mono.Linker.Tests.Cases.CommandLine.Mvid
+namespace Mono.Linker.Tests.Cases.CommandLine.Mvid.Individual
 {
-    [SetupLinkerArgument("--deterministic", "true")]
-    public class DeterministicMvidWorks
+    [SetupLinkerArgument("--new-mvid", "true")]
+    public class NewMvidWorks
     {
         public static void Main()
         {
