+++ src/tools/illink/test/Mono.Linker.Tests.Cases/CommandLine/Mvid/Individual/RetainMvid.cs
@@ -1,10 +1,10 @@
 using Mono.Linker.Tests.Cases.Expectations.Assertions;
 using Mono.Linker.Tests.Cases.Expectations.Metadata;
 
-namespace Mono.Linker.Tests.Cases.CommandLine.Mvid
+namespace Mono.Linker.Tests.Cases.CommandLine.Mvid.Individual
 {
-    [SetupLinkerArgument("--new-mvid", "true")]
-    public class NewMvidWorks
+    [SetupLinkerArgument("--new-mvid", "false")]
+    public class RetainMvid
     {
         public static void Main()
         {
